refactor(options): split Stitch schema renderer
