+++ scripts/preview-freeze-check.mjs
@@ -52,9 +52,9 @@ const ALLOWED_PREVIEW_DRIFT = new Map([
     'original reference index.js vs current generated preview index.js',
     {
       reason:
-        'Generated Stitch Secondary preview JS intentionally carries the previously approved SVG-aware Stitch DOM helper, production parity/runtime harness wiring, schema-registered runtime surfaces, Reader/Video item-level session controls, the shared Reader/Video collapse action, Clipper removal of preview-incompatible visible cancel/comment/selection labels and the empty comment preview block, the inline video capture action plus readonly add-note trigger, persistent video control-bar auto-pause/screenshot preferences, the System interface theme preference, shared Clipper/Reader/Video export destination preview/dropdown content without visible destination carets, shared session width/height resize handle wiring, video fragment captures rendered as reader-style numbered text items with text-note placeholders, All in Ob production branding, QR-free task-success feedback toasts, the product-approved removal of the visible Settings sidebar title, production onboarding page routing from the Options resource link, release-hidden Experimental panel, Deep Research, AI timestamp, advanced video schema, prompt label/shortcut controls, and classifier placeholders, the performance branch classification timeout fallback enum, one-month Usage Dashboard history padding, Fragment modifier in-place chip sync, YAML button event containment, mouse-scroll-stable Stitch action buttons, render scroll restoration for Options actions, the video note-button switch promoted into the Video Prompt & Entry header with gray/green screenshot-dot guidance, default Domain Mappings restoration with an editable blank row only after user deletion plus inline-scrolling table chrome, Routing Rules canonical de-duplication, video timestamp screenshot toggle state rendered before the timestamp label, and the product-approved Reader/Video/task-success bottom-right floating runtime contract with no static task-success toast previews, compact feedback row, and Stitch-tokenized runtime toast root. Batch 2.5 additionally adds REST and vault local-folder File System Access controls, local-folder picker/remove/reauthorization actions, default-vault local-folder metadata syncing, storage schema wiring for local-folder rows, Options runtime routing/bootstrap preservation, button-action scroll preservation, and a task-success progress schema/content field absent from the frozen original reference. Milestone 7 relocates the preview-backed verification runtime from the retired source tree to tests/fixtures/options-preview while preserving the same frozen reference comparison contract. Milestone 10 splits the YAML Config Widget into adapter, controller, model conversion, validation presentation, field table rendering, and domain override rendering modules while preserving the production Stitch runtime contract. Milestone 13 splits the Stitch renderer into action, form, content, and node renderer modules while preserving the production Stitch runtime contract. Final verification stabilizes the preview bundle hash after the FragmentSectionView collectChanges return simplification without changing the frozen reference, freeze-check logic, or production behavior. The completion-audit format pass refreshes the generated JS hash after repo-wide Prettier normalization without changing the frozen reference or preview behavior.',
+        'Generated Stitch Secondary preview JS intentionally carries the previously approved SVG-aware Stitch DOM helper, production parity/runtime harness wiring, schema-registered runtime surfaces, Reader/Video item-level session controls, the shared Reader/Video collapse action, Clipper removal of preview-incompatible visible cancel/comment/selection labels and the empty comment preview block, the inline video capture action plus readonly add-note trigger, persistent video control-bar auto-pause/screenshot preferences, the System interface theme preference, shared Clipper/Reader/Video export destination preview/dropdown content without visible destination carets, shared session width/height resize handle wiring, video fragment captures rendered as reader-style numbered text items with text-note placeholders, All in Ob production branding, QR-free task-success feedback toasts, the product-approved removal of the visible Settings sidebar title, production onboarding page routing from the Options resource link, release-hidden Experimental panel, Deep Research, AI timestamp, advanced video schema, prompt label/shortcut controls, and classifier placeholders, the performance branch classification timeout fallback enum, one-month Usage Dashboard history padding, Fragment modifier in-place chip sync, YAML button event containment, mouse-scroll-stable Stitch action buttons, render scroll restoration for Options actions, the video note-button switch promoted into the Video Prompt & Entry header with gray/green screenshot-dot guidance, default Domain Mappings restoration with an editable blank row only after user deletion plus inline-scrolling table chrome, Routing Rules canonical de-duplication, video timestamp screenshot toggle state rendered before the timestamp label, and the product-approved Reader/Video/task-success bottom-right floating runtime contract with no static task-success toast previews, compact feedback row, and Stitch-tokenized runtime toast root. Batch 2.5 additionally adds REST and vault local-folder File System Access controls, local-folder picker/remove/reauthorization actions, default-vault local-folder metadata syncing, storage schema wiring for local-folder rows, Options runtime routing/bootstrap preservation, button-action scroll preservation, and a task-success progress schema/content field absent from the frozen original reference. Milestone 7 relocates the preview-backed verification runtime from the retired source tree to tests/fixtures/options-preview while preserving the same frozen reference comparison contract. Milestone 10 splits the YAML Config Widget into adapter, controller, model conversion, validation presentation, field table rendering, and domain override rendering modules while preserving the production Stitch runtime contract. Milestone 13 splits the Stitch renderer into action, form, content, and node renderer modules while preserving the production Stitch runtime contract. Final verification stabilizes the preview bundle hash after the FragmentSectionView collectChanges return simplification without changing the frozen reference, freeze-check logic, or production behavior. The completion-audit format pass refreshes the generated JS hash after repo-wide Prettier normalization without changing the frozen reference or preview behavior. M3.1 splits the schema runtime renderer into context, widget lifecycle, hero, control, table, structural, element, node dispatch, and view modules while preserving the frozen reference and production Stitch runtime behavior.',
       leftSha256: '9020ccbd91acd691eccd3fdf568b9a90efbddf0a35d79f36ef1caba702fa0c07',
-      rightSha256: 'c5fd69c90f86f45a7f832896666e891efe6a2240a50fd9a75c92441fd1b4ee22'
+      rightSha256: '7e9c3d6b30ad451a08a09892de75a1d73d85c78826e9eb6e0efb0f35da75c223'
     }
   ]
 ]);
